@@ -16,14 +16,16 @@
         "inertiajs/inertia-laravel": "^0.4.2",
         "itsgoingd/clockwork": "^5.0",
         "laravel/fortify": "^1.7",
-        "laravel/framework": "^8.12",
+        "laravel/framework": "8.48.1",
         "laravel/sanctum": "^2.11",
         "laravel/tinker": "^2.5",
         "livewire/livewire": "^2.4",
         "maatwebsite/excel": "^3.1",
+        "spatie/eloquent-sortable": "^4.0",
         "spatie/laravel-enum": "^2.5",
         "spatie/laravel-query-builder": "^3.4",
         "spatie/laravel-route-attributes": "^1.4",
+        "spatie/laravel-sluggable": "^3.1",
         "tightenco/ziggy": "^1.1"
     },
     "require-dev": {
@@ -36,7 +38,8 @@
         "nunomaduro/collision": "^5.0",
         "nunomaduro/larastan": "^0.7.6",
         "pestphp/pest-plugin-laravel": "^1.0",
-        "phpunit/phpunit": "^9.3.3"
+        "phpunit/phpunit": "^9.3.3",
+        "spatie/phpunit-watcher": "^1.23"
     },
     "autoload": {
         "psr-4": {
@@ -74,6 +77,9 @@
         ],
         "test": [
             "@php artisan test --coverage --min=90"
+        ],
+        "test:watch": [
+            "phpunit-watcher watch"
         ]
     },
     "extra": {
