@@ -1,49 +1,71 @@
 <template>
-  <component
-    :is="href ? tag : 'button'"
-    class="
-      inline-flex
-      justify-center
-      items-center
-      px-4
-      py-2
-      border border-transparent
-      rounded-md
-      font-semibold
-      text-xs
-      uppercase
-      tracking-widest
-      focus:outline-none focus:ring
-      disabled:opacity-25
-      transition
-    "
-    :class="[
-      `btn-${variant}`,
-      { 'btn-outlined': outlined, 'opacity-25': loading },
-    ]"
-    :disabled="loading"
-    :href="href"
-    :only="only"
-  >
-    <spinner
-      v-if="loading"
-      class="-ml-1 h-4 w-4 text-white"
-      :class="{
-        '-ml-1 mr-2': hasLabel,
-        '-ml-1 -mr-2': !hasLabel,
-      }"
-    />
+  <span class="inline-flex">
     <component
-      :is="`${icon}-icon`"
-      v-else-if="icon"
-      class="w-4 h-4"
-      :class="{
-        '-ml-1 mr-2': hasLabel,
-        '-ml-2 -mr-2': !hasLabel,
-      }"
-    />
-    <slot v-if="hasLabel"></slot>
-  </component>
+      v-bind="$attrs"
+      :is="href ? tag : 'button'"
+      class="
+        inline-flex
+        justify-center
+        items-center
+        px-4
+        py-2
+        border border-transparent
+        rounded-l-md
+        text-xs
+        uppercase
+        font-semibold
+        tracking-widest
+        focus:outline-none focus:ring
+        disabled:opacity-25
+        transition
+      "
+      :class="[
+        `btn-${variant}`,
+        {
+          'btn-outlined': outlined,
+          'opacity-25': loading,
+          'rounded-r-md': !split,
+        },
+      ]"
+      :disabled="loading"
+      :href="href"
+      :only="only"
+      @click.stop="$emit('click')"
+    >
+      <spinner
+        v-if="loading"
+        class="-ml-1 h-4 w-4 text-white"
+        :class="{
+          '-ml-1 mr-2': hasLabel,
+          '-ml-1 -mr-2': !hasLabel,
+        }"
+      />
+      <component
+        :is="`${icon}-icon`"
+        v-else-if="icon"
+        class="w-4 h-4"
+        :class="{
+          '-ml-1 mr-2': hasLabel,
+          '-ml-2 -mr-2': !hasLabel,
+        }"
+      />
+      <slot v-if="hasLabel"></slot>
+    </component>
+    <button
+      v-if="split"
+      type="button"
+      class="px-2 py-2 border border-transparent rounded-r-md"
+      :class="[
+        `btn-${variant}`,
+        {
+          'btn-outlined': outlined,
+          'opacity-25': loading,
+        },
+      ]"
+    >
+      <chevron-down-icon class="h-4 w-4" />
+    </button>
+  </span>
 </template>
 
 <script lang="ts">
@@ -56,6 +78,7 @@
       href: String,
       icon: String,
       hideLabel: Boolean,
+      split: Boolean,
       tag: {
         type: String,
         default: 'inertia-link',
@@ -70,6 +93,7 @@
       },
       only: Array,
     },
+    emits: ['click'],
     setup(props, { slots }) {
       const hasLabel = computed(() => {
         return !props.hideLabel && !!slots.default
