@@ -3,5 +3,6 @@ import { EnumTypes } from './enums'
 import { Model } from './model'
 import { Profile } from './profile'
 import { User } from './user'
+import { Post } from './post'
 
-export { PaginatedData, EnumTypes, Model, Profile, User }
+export { PaginatedData, EnumTypes, Model, Profile, User, Post }
